adding visuals for eq, blanking, and matrix subpages.
I have completed the eq section of the program however I have only added the blanking and matrix subpages but not added the functionality that needs to exist for those two pages. I have imported the smart objects in to the simpl windows program, there is lots to come for the simpl windows programming however that is not happening today.