Adding matrix support, fully working
The matrix has been redesigned to use vertical smart objects and now is setup to properly configure analog outputs. This now also includes program audio, which will hide out the buttons (in program audio) that are in accessible.

TODO: make the button enable more granular so that if one display is using laptop and another is using desktop it shows the respective labels. i might also need to restrict the matrix buttons based on which transmitter is being used as some combinations are not possible since the transmitter is only able to show one output at a given time.